# sacdia-api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente"
          },
          "400": {
            "description": "Error en validación o usuario ya existe"
          }
        },
        "summary": "Registrar nuevo usuario",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso, retorna tokens y usuario"
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sesión cerrada exitosamente"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cerrar sesión",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password/reset-request": {
      "post": {
        "operationId": "AuthController_requestPasswordReset_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Correo de recuperación enviado"
          }
        },
        "summary": "Solicitar recuperación de contraseña",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Información del usuario"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener perfil del usuario autenticado",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/profile/completion-status": {
      "get": {
        "operationId": "AuthController_getCompletionStatus_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Estado del post-registro del usuario"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener estado del post-registro",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/clubs/{clubId}/activities": {
      "get": {
        "description": "Obtiene todas las actividades de las instancias del club",
        "operationId": "ActivitiesController_findByClub_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "clubTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "activityType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de actividades"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar actividades del club",
        "tags": [
          "activities"
        ]
      },
      "post": {
        "description": "Crea una nueva actividad para el club (requiere rol de liderazgo)",
        "operationId": "ActivitiesController_create_v1",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivityDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Actividad creada"
          },
          "403": {
            "description": "Permisos insuficientes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear actividad",
        "tags": [
          "activities"
        ]
      }
    },
    "/v1/activities/{activityId}": {
      "get": {
        "operationId": "ActivitiesController_findOne_v1",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Actividad encontrada"
          },
          "404": {
            "description": "Actividad no encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener actividad por ID",
        "tags": [
          "activities"
        ]
      },
      "patch": {
        "operationId": "ActivitiesController_update_v1",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateActivityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actividad actualizada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar actividad",
        "tags": [
          "activities"
        ]
      },
      "delete": {
        "operationId": "ActivitiesController_remove_v1",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Actividad desactivada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Desactivar actividad",
        "tags": [
          "activities"
        ]
      }
    },
    "/v1/activities/{activityId}/attendance": {
      "post": {
        "description": "Registra la lista de usuarios que asistieron a la actividad",
        "operationId": "ActivitiesController_recordAttendance_v1",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Asistencia registrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar asistencia",
        "tags": [
          "activities"
        ]
      },
      "get": {
        "description": "Obtiene la lista de usuarios que asistieron a la actividad",
        "operationId": "ActivitiesController_getAttendance_v1",
        "parameters": [
          {
            "name": "activityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de asistentes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener asistencia",
        "tags": [
          "activities"
        ]
      }
    },
...
  },
  "info": {
    "title": "SACDIA API",
    "description": "## Sistema de Administración de Clubes de Conquistadores y Aventureros\n\n### Módulos Disponibles\n- **Auth**: Autenticación con Supabase + JWT\n- **Users**: Gestión de perfiles de usuario\n- **Catalogs**: Catálogos de referencia\n- **Clubs**: Gestión de clubes e instancias\n- **Classes**: Clases progresivas\n- **Honors**: Especialidades\n- **Activities**: Actividades de club\n- **Finances**: Control financiero",
    "version": "2.2.0",
    "contact": {
      "name": "SACDIA Team",
      "url": "https://sacdia.app",
      "email": "dev@sacdia.app"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "example": "Juan",
            "description": "Nombre del usuario"
          },
          "paternal_last_name": {
            "type": "string",
            "maxLength": 50,
            "example": "García",
            "description": "Apellido paterno"
          },
          "maternal_last_name": {
            "type": "string",
            "maxLength": 50,
            "example": "López",
            "description": "Apellido materno"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "juan.garcia@example.com",
            "description": "Correo electrónico"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "pattern": "/^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]/",
            "example": "Password123!",
            "description": "Contraseña segura (mínimo 8 caracteres, debe incluir mayúscula, minúscula, número y carácter especial)"
          }
        },
        "required": [
          "name",
          "paternal_last_name",
          "maternal_last_name",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "juan.garcia@example.com",
            "description": "Correo electrónico"
          },
          "password": {
            "type": "string",
            "example": "Password123!",
            "description": "Contraseña"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ResetPasswordRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "juan.garcia@example.com",
            "description": "Correo para recuperación"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "CreateActivityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre de la actividad"
          },
          "description": {
            "type": "string",
            "description": "Descripción de la actividad"
          },
          "club_type_id": {
            "type": "number",
            "description": "Tipo de club (1=Aventureros, 2=Conquistadores, 3=GM)"
          },
          "lat": {
            "type": "number",
            "minimum": -90,
            "maximum": 90,
            "description": "Latitud del lugar"
          },
          "long": {
            "type": "number",
            "minimum": -180,
            "maximum": 180,
            "description": "Longitud del lugar"
          },
          "activity_time": {
            "type": "string",
            "description": "Hora de la actividad (HH:mm)",
            "default": "09:00"
          },
          "activity_place": {
            "type": "string",
            "description": "Lugar de la actividad"
          },
          "image": {
            "type": "string",
            "description": "URL de la imagen de la actividad"
          },
          "platform": {
            "type": "number",
            "description": "Plataforma (0=Presencial, 1=Virtual, 2=Híbrido)",
            "default": 0
          },
          "activity_type": {
            "type": "number",
            "description": "Tipo de actividad (0=Regular, 1=Especial, 2=Camporee)",
            "default": 0
          },
          "link_meet": {
            "type": "string",
            "description": "Link de reunión virtual"
          },
          "additional_data": {
            "type": "string",
            "description": "Datos adicionales en JSON"
          },
          "classes": {
            "description": "Clases invitadas (IDs)",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "club_adv_id": {
            "type": "number",
            "description": "ID de la instancia de Aventureros"
          },
          "club_pathf_id": {
            "type": "number",
            "description": "ID de la instancia de Conquistadores"
          },
          "club_mg_id": {
            "type": "number",
            "description": "ID de la instancia de Guías Mayores"
          }
        },
        "required": [
          "name",
          "club_type_id",
          "lat",
          "long",
          "activity_place",
          "image",
          "club_adv_id",
          "club_pathf_id",
          "club_mg_id"
        ]
      },
      "UpdateActivityDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "lat": {
            "type": "number"
          },
          "long": {
            "type": "number"
          },
          "activity_time": {
            "type": "string"
          },
          "activity_place": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "platform": {
            "type": "number"
          },
          "activity_type": {
            "type": "number"
          },
          "link_meet": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "classes": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
      "RecordAttendanceDto": {
        "type": "object",
        "properties": {
          "user_ids": {
            "description": "Lista de IDs de usuarios que asistieron",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "user_ids"
        ]
      },
...
    }
  }
}